feat(workout-builder): add ExercisesSelection and WorkoutBuilderFooter components to enhance workout selection and navigation experience refactor(workout-stepper): simplify workout stepper by integrating new components and removing redundant code for better maintainability
